+++ mpp/mpp_utilities.F90
@@ -59,11 +59,11 @@ subroutine mpp_array_global_min_max(in_array, tmask,isd,jsd,isc,iec,jsc,jec,nk, 
   real,                         intent(out):: g_min, g_max
   real, dimension(isd:,jsd:),   intent(in) :: geo_x,geo_y
   real, dimension(:),           intent(in) :: geo_z
-  real,                         intent(out):: xgmin, ygmin, zgmin, xgmax, ygmax, zgmax 
+  real(kind=8),                 intent(out):: xgmin, ygmin, zgmin, xgmax, ygmax, zgmax
 
 
 
-  real    :: tmax, tmin, tmax0, tmin0
+  real(kind=8)    :: tmax, tmin, tmax0, tmin0
   integer :: itmax, jtmax, ktmax, itmin, jtmin, ktmin
   integer :: igmax, jgmax, kgmax, igmin, jgmin, kgmin
   real    :: fudge
